x86-64向けにもビルドできるようにする
ほとんどがcchSizeのUINT→size_t化と
Set(Window|Class)LongPtrの修正

SIZE_Tとsize_tは微妙に違うみたいで大文字のSIZE_Tを使うとi686向けのコンパイルが壊れた
そのため小文字のsize_tを使ってる

一応、一通りのビルドと起動の確認はできたが細かいところは未確認

+++ src/DocAdjustant.cpp
@@ -874,7 +874,7 @@ LPTSTR DocPaddingSpaceWithPeriod( INT dTgtDot, PINT pdZen, PINT pdHan, PINT pdPr
 {
 	INT	dZenSp, dHanSp, dPrdSp, m, dPre;
 	LPTSTR	ptSpace = NULL, ptPlus = NULL;
-	UINT	cchSize, cchPlus;
+	size_t	cchSize, cchPlus;
 
 
 	dPre = dTgtDot;
